requirements.txt
tensorflow>=2.8.0,<2.20.0
flask>=2.0.0
opencv-python>=4.5.0
numpy>=1.23.0
pillow>=8.0.0
scikit-learn>=0.24.0
werkzeug>=2.0.0
python-dotenv>=0.19.0
gunicorn>=20.1.0 